Extract task reminder controller slice
