+++ apps/core-api/src/app.module.ts
@@ -44,6 +44,8 @@ import { TaskAttachmentsService } from './tasks/task-attachments.service';
 import { TaskCommentsController } from './tasks/task-comments.controller';
 import { TaskCommentsService } from './tasks/task-comments.service';
 import { TaskMentionsService } from './tasks/task-mentions.service';
+import { TaskRemindersController } from './tasks/task-reminders.controller';
+import { TaskRemindersService } from './tasks/task-reminders.service';
 import { GoalsModule } from './goals/goals.module';
 import { CapacityModule } from './capacity/capacity.module';
 import { GuestAccessController } from './guest-access/guest-access.controller';
@@ -74,6 +76,7 @@ import { GuestAccessController } from './guest-access/guest-access.controller';
     PublicAttachmentsController,
     TaskAttachmentsController,
     TaskCommentsController,
+    TaskRemindersController,
     CollabController,
     WorkspaceAdminController,
     NotificationsController,
@@ -99,6 +102,7 @@ import { GuestAccessController } from './guest-access/guest-access.controller';
     TaskAttachmentsService,
     TaskCommentsService,
     TaskMentionsService,
+    TaskRemindersService,
     PublicFormSubmissionThrottleGuard,
     ProjectRoleGuard,
     WorkspaceRoleGuard,
